requirements.txt
djangorestframework==3.14.0
drf-yasg==1.21.7
Django==4.2.4
